[수정] 상품환불시 상품/재구매 마다 환불포인트 분기 및 최대수당제한 차감

+++ adm/shop_admin/order_proc.php
@@ -19,41 +19,56 @@ function  od_name_return_rank($val){
 
 if($od_item){
 
-    //상품생성 테이블 삭제 
-    $rank_num = od_name_return_rank($od_item['od_name']);
-    $package_group = "package_p".$rank_num;
-    $package_have_sql = "SELECT * from {$package_group} WHERE od_id = '{$od_item['od_id']}' ";
-    $package_have = sql_fetch($package_have_sql);
-
-    print_R($package_have_sql);
-    echo "<br><br>";
-
-    if($package_have){
-        $del_package = "DELETE FROM {$package_group} WHERE od_id = '{$od_item['od_id']}' ";
-
-        print_R($del_package);
-        echo "<br><br>";
-
-        $pack_del_result = sql_query($del_package);
-    }
-
-    // 금액반환처리
     $amt = $od_item['od_cart_price'];
     $upstair = $od_item['od_cash'];
     $pv = $od_item['pv'];
-    $update_member_sql = "UPDATE g5_member set mb_deposit_calc= mb_deposit_calc + {$amt}, mb_save_point = mb_save_point - {$upstair}, mb_rate = mb_rate - {$pv} ";
+    $od_misu = $od_item['od_misu'];
+    $mb_id = $od_item['mb_id'];
 
-    if($rank_num == 0){
-        $update_member_sql .=", sales_day = '0000-00-00' , rank_note = '' , rank = 0 ";
+    if($od_item['od_cash_no'] != "재구매"){
+
+        //상품생성 테이블 삭제 
+        $rank_num = od_name_return_rank($od_item['od_name']);
+        $package_group = "package_p".$rank_num;
+        $package_have_sql = "SELECT * from {$package_group} WHERE od_id = '{$od_item['od_id']}' ";
+        $package_have = sql_fetch($package_have_sql);
+    
+        print_R($package_have_sql);
+        echo "<br><br>";
+    
+        if($package_have){
+            $del_package = "DELETE FROM {$package_group} WHERE od_id = '{$od_item['od_id']}' ";
+    
+            print_R($del_package);
+            echo "<br><br>";
+    
+            $pack_del_result = sql_query($del_package);
+        }
+    
+        // 금액반환처리
+
+        $update_member_sql = "UPDATE g5_member set mb_deposit_calc= mb_deposit_calc + {$amt}, mb_save_point = mb_save_point - {$upstair}, mb_rate = mb_rate - {$pv}, 
+        pv = pv - {$amt}, mb_index = mb_index - {$od_misu} ";
+    
+        if($rank_num == 0){
+            $update_member_sql .=", sales_day = '0000-00-00' , rank_note = '' , rank = 0 ";
+        }else{
+            $update_member_sql .=", rank_note = 'Membership-PACK' , rank = 1 ";
+        }
+        $update_member_sql .= " WHERE mb_id = '{$mb_id}' ";
+    
+        print_R($update_member_sql);
+        echo "<br><br>";
+        // $update_result = 1;
+        $update_result = sql_query($update_member_sql);
+    
+
     }else{
-        $update_member_sql .=", rank_note = 'Membership-PACK' , rank = 1 ";
+        $update_member_sql = "UPDATE g5_member SET pv = pv - {$amt}, 
+        mb_balance_calc = mb_balance_calc + {$amt}, mb_index = mb_index - {$od_misu} WHERE mb_id = '{$mb_id}'";
+    
+        $update_result = sql_query($update_member_sql);
     }
-    $update_member_sql .= " WHERE mb_id = '{$od_item['mb_id']}' ";
-
-    print_R($update_member_sql);
-    echo "<br><br>";
-    // $update_result = 1;
-    $update_result = sql_query($update_member_sql);
 
     if($update_result){
         $de_data = $od_item['od_name']." | ".$amt." | ".$od_item['od_status'].' 건 구매취소처리';
